backend/requirements.txt
# FastAPI & server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
supabase==2.7.4
asyncpg==0.31.0
sqlalchemy>=2.0.35
alembic>=1.13.3

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.27.2

# Workers
celery>=5.4.0
redis>=5.1.0
flower>=2.0.1

# Stripe
stripe>=10.12.0

# Email
resend>=2.3.0

# Monitoring utilities
dnspython>=2.7.0
cryptography>=43.0.1
pyOpenSSL>=24.2.1
requests>=2.32.3

# AI
anthropic>=0.34.2

# Config & utils
pydantic>=2.9.2
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
APScheduler>=3.10.4
tenacity>=9.0.0
structlog>=24.4.0
